+++ NEWS.md
@@ -2,7 +2,35 @@
 
 ## Changes
 
+- `plot()` for `p_function()` now checks the values of the `size_length`
+  argument, to give an informative error message when the input is not valid.
+
+- `plot()` for `model_parameters()` now also plots group-levels of random effects
+  (i.e. for mixed models, when `model_parameters(x, ..., group_level = TRUE)`).
+
+- `plot()` for `model_parameters()` gets a `show_direction` argument, to turn
+  off the direction of the effect in the plot.
+
+- `plot()` for `simulate_parameters()` now better copes with models that have
+  multiple response levels (e.g. multinomial models).
+
+## Bug fixes
+
+- Fixed issue in `plot()` for `parameters::model_parameters()` for GAM models.
+
+# see 0.9.0
+
+## Changes
+
 - New `plot()` method for `performance::check_dag()`.
+
+- Minor improvements to `plot()` for methods `p_direction()` and `p_significance()`,
+  which also support forthcoming changes in the *parameters* package.
+
+## Bug fixes
+
+- Fixed issue in `plot()` for `performance::check_model()` when package *qqplotr*
+  is not installed.
 
 # see 0.8.5